Refactoring the core logic from sendgrid controller to a separate service
