+++ Theatre_Timeline/Controllers/SendGridController.cs
@@ -1,7 +1,5 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Text.Json;
-using System.Text.RegularExpressions;
 using Theatre_TimeLine.Contracts;
 using Theatre_TimeLine.Models;
 using Theatre_TimeLine.Services;
@@ -13,55 +11,20 @@ namespace Theatre_TimeLine.Controllers
     /// </summary>
     [ApiController]
     [Route("api/[controller]")]
-    public partial class SendGridController : ControllerBase
+    public class SendGridController : ControllerBase
     {
-        // Regex to sanitize log input - removes newlines and control characters to prevent log injection
-        [GeneratedRegex(@"[\r\n\t\x00-\x1F\x7F]", RegexOptions.Compiled)]
-        private static partial Regex LogSanitizationRegex();
-
         private readonly ILogger<SendGridController> _logger;
-        private readonly IConfiguration _configuration;
-        private readonly IEmailEncryptionService _encryptionService;
         private readonly ISecurityGroupService _securityGroupService;
-        private readonly ISendGridWebhookValidator _webhookValidator;
-        private readonly string _emailStoragePath;
+        private readonly ISendGridEmailService _sendGridEmailService;
 
         public SendGridController(
             ILogger<SendGridController> logger,
-            IConfiguration configuration,
-            IEmailEncryptionService encryptionService,
             ISecurityGroupService securityGroupService,
-            ISendGridWebhookValidator webhookValidator)
+            ISendGridEmailService sendGridEmailService)
         {
             _logger = logger;
-            _configuration = configuration;
-            _encryptionService = encryptionService;
             _securityGroupService = securityGroupService;
-            _webhookValidator = webhookValidator;
-
-            // Get storage path from configuration or use default
-            string? emailPath = _configuration.GetValue<string>("SendGrid:EmailStoragePath");
-
-            if (string.IsNullOrEmpty(emailPath))
-            {
-                emailPath = "./emails";
-            }
-
-            if (!Path.IsPathRooted(emailPath))
-            {
-                emailPath = Path.Combine(
-                    AppDomain.CurrentDomain.BaseDirectory,
-                    emailPath.Trim(['.', '\\', '/']));
-            }
-
-            _emailStoragePath = emailPath;
-
-            // Ensure directory exists
-            if (!Directory.Exists(_emailStoragePath))
-            {
-                Directory.CreateDirectory(_emailStoragePath);
-                _logger.LogInformation("Created email storage directory: {Path}", _emailStoragePath);
-            }
+            _sendGridEmailService = sendGridEmailService;
         }
 
         /// <summary>
@@ -79,84 +42,21 @@ namespace Theatre_TimeLine.Controllers
         {
             try
             {
-                _logger.LogInformation("Received inbound email request from SendGrid");
+                var result = await _sendGridEmailService.ProcessInboundEmailAsync(HttpContext, email);
 
-                // Validate the webhook source
-                var validationResult = await _webhookValidator.ValidateRequestAsync(HttpContext);
-
-                if (!validationResult.IsValid)
+                if (!result.IsValid)
                 {
-                    _logger.LogWarning("Webhook validation failed: {Reason}", SanitizeForLog(validationResult.Reason));
-                    return Unauthorized(new { error = "Invalid webhook source", reason = validationResult.Reason });
+                    return Unauthorized(new { error = "Invalid webhook source", reason = result.ValidationReason });
                 }
-
-                _logger.LogInformation("Webhook validation passed. Headers captured: {HasHeaders}",
-                    validationResult.Headers != null);
-
-                // Set metadata fields (not bound from form)
-                email.ReceivedAt = DateTime.UtcNow.ToString("O");
-                email.Encrypted = true;
-                email.ValidatedSource = validationResult.IsValid;
-                email.WebhookHeaders = validationResult.Headers;
-
-                // Handle file attachments if present
-                var form = await Request.ReadFormAsync();
-                if (form.Files.Count > 0)
-                {
-                    email.Attachments = form.Files.Select(file => new EmailAttachment
-                    {
-                        Filename = file.FileName,
-                        ContentType = file.ContentType,
-                        Length = file.Length
-                    }).ToList();
-                }
-
-                // Log parsed email info (sanitized and masked to protect PII)
-                _logger.LogInformation("Parsed email - From: {From}, To: {To}, Subject: {Subject}",
-                    MaskEmail(email.GetFromEmail()),
-                    MaskEmail(email.GetToEmail()),
-                    SanitizeForLog(email.Subject));
-                _logger.LogInformation("Email validation - DKIM: {Dkim}, SPF: {Spf}, Spam Score: {SpamScore}",
-                    email.IsDkimValid(), email.IsSpfValid(), email.GetSpamScoreValue());
-
-                // Check if email is spam
-                if (email.IsLikelySpam())
-                {
-                    _logger.LogWarning("Email flagged as spam (score: {SpamScore}), saving but marking as spam",
-                        email.GetSpamScoreValue());
-                }
-
-                // Generate filename with timestamp
-                string timestamp = DateTime.UtcNow.ToString("yyyyMMdd_HHmmss");
-                string fromEmail = email.GetFromEmail()?.Replace("@", "_at_") ?? "unknown";
-                string sanitizedFrom = string.Join("_", fromEmail.Split(Path.GetInvalidFileNameChars()));
-                string filename = $"email_{timestamp}_{sanitizedFrom}_{Guid.NewGuid()}.enc";
-                string filePath = Path.Combine(_emailStoragePath, filename);
-
-                // Store the filename in the model
-                email.StoredAs = filename;
-
-                // Serialize to JSON
-                var options = new JsonSerializerOptions
-                {
-                    WriteIndented = true,
-                    DefaultIgnoreCondition = System.Text.Json.Serialization.JsonIgnoreCondition.WhenWritingNull
-                };
-                string jsonContent = JsonSerializer.Serialize(email, options);
-
-                // Encrypt and write to disk
-                await _encryptionService.WriteEncryptedFileAsync(filePath, jsonContent);
-
-                _logger.LogInformation("Encrypted email saved successfully");
 
                 return Ok(new
                 {
                     message = "Email received, validated, encrypted, and saved successfully",
-                    file = filename,
-                    from = email.GetFromEmail(),
-                    subject = email.Subject,
-                    spamScore = email.GetSpamScoreValue(),
-                    validated = validationResult.IsValid
+                    file = result.Filename,
+                    from = result.From,
+                    subject = result.Subject,
+                    spamScore = result.SpamScore,
+                    validated = result.IsValid
                 });
             }
             catch (Exception ex)
@@ -177,7 +77,6 @@ namespace Theatre_TimeLine.Controllers
         [Authorize]
         public async Task<IActionResult> GetEmail(string filename)
         {
-            // Check if user is Global Admin
             if (!await IsGlobalAdminAsync())
             {
                 _logger.LogWarning("Unauthorized access attempt to email by user");
@@ -186,55 +85,29 @@ namespace Theatre_TimeLine.Controllers
 
             try
             {
-                // Sanitize filename to prevent directory traversal
-                var safeFilename = Path.GetFileName(filename);
+                var result = await _sendGridEmailService.GetStoredEmailAsync(filename);
 
-                // Additional validation: ensure filename matches expected pattern
-                if (string.IsNullOrEmpty(safeFilename) ||
-                    !safeFilename.StartsWith("email_", StringComparison.Ordinal) ||
-                    !safeFilename.EndsWith(".enc", StringComparison.Ordinal))
+                return result.Status switch
                 {
-                    return BadRequest(new { error = "Invalid filename format" });
-                }
-
-                var filePath = Path.Combine(_emailStoragePath, safeFilename);
-
-                // Verify the resolved path is within the allowed directory (defense in depth)
-                var fullPath = Path.GetFullPath(filePath);
-                var allowedPath = Path.GetFullPath(_emailStoragePath);
-                if (!fullPath.StartsWith(allowedPath, StringComparison.OrdinalIgnoreCase))
-                {
-                    _logger.LogWarning("Path traversal attempt detected");
-                    return BadRequest(new { error = "Invalid path" });
-                }
-
-                if (!System.IO.File.Exists(fullPath))
-                {
-                    return NotFound(new { error = "Email file not found" });
-                }
-
-                // Read and decrypt
-                var decryptedContent = await _encryptionService.ReadEncryptedFileAsync(fullPath);
-                var email = JsonSerializer.Deserialize<SendGridInboundEmail>(decryptedContent);
-
-                _logger.LogInformation("Retrieved and decrypted email successfully");
-
-                // Return with parsed body content for convenience
-                return Ok(new
-                {
-                    email,
-                    parsed = new
+                    StoredEmailLookupStatus.InvalidFilename => BadRequest(new { error = "Invalid filename format" }),
+                    StoredEmailLookupStatus.InvalidPath => BadRequest(new { error = "Invalid path" }),
+                    StoredEmailLookupStatus.NotFound => NotFound(new { error = "Email file not found" }),
+                    _ => Ok(new
                     {
-                        fromEmail = email?.GetFromEmail(),
-                        fromName = email?.GetFromDisplayName(),
-                        toEmail = email?.GetToEmail(),
-                        bodyText = email?.GetTextBody(),
-                        bodyContent = email?.GetBodyContent(),
-                        isDkimValid = email?.IsDkimValid(),
-                        isSpfValid = email?.IsSpfValid(),
-                        isSpam = email?.IsLikelySpam()
-                    }
-                });
+                        email = result.Email,
+                        parsed = new
+                        {
+                            fromEmail = result.Email?.GetFromEmail(),
+                            fromName = result.Email?.GetFromDisplayName(),
+                            toEmail = result.Email?.GetToEmail(),
+                            bodyText = result.Email?.GetTextBody(),
+                            bodyContent = result.Email?.GetBodyContent(),
+                            isDkimValid = result.Email?.IsDkimValid(),
+                            isSpfValid = result.Email?.IsSpfValid(),
+                            isSpam = result.Email?.IsLikelySpam()
+                        }
+                    })
+                };
             }
             catch (Exception ex)
             {
@@ -253,7 +126,6 @@ namespace Theatre_TimeLine.Controllers
         [Authorize]
         public async Task<IActionResult> ListEmails()
         {
-            // Check if user is Global Admin
             if (!await IsGlobalAdminAsync())
             {
                 _logger.LogWarning("Unauthorized access attempt to email list");
@@ -262,16 +134,7 @@ namespace Theatre_TimeLine.Controllers
 
             try
             {
-                var files = Directory.GetFiles(_emailStoragePath, "email_*.enc")
-                    .Select(f => new
-                    {
-                        filename = Path.GetFileName(f),
-                        size = new FileInfo(f).Length,
-                        created = System.IO.File.GetCreationTimeUtc(f)
-                    })
-                    .OrderByDescending(f => f.created)
-                    .ToList();
-
+                var files = await _sendGridEmailService.ListStoredEmailsAsync();
                 return Ok(new { count = files.Count, emails = files });
             }
             catch (Exception ex)
@@ -289,96 +152,7 @@ namespace Theatre_TimeLine.Controllers
         [AllowAnonymous]
         public IActionResult HealthCheck()
         {
-            // Get encryption status from configuration
-            bool encryptionEnabled = _configuration.GetValue<bool>("SendGrid:EnableEncryption", true);
-            bool requireIpValidation = _configuration.GetValue<bool>("SendGrid:RequireIpValidation", false);
-            bool requireAuthValidation = _configuration.GetValue<bool>("SendGrid:RequireAuthValidation", false);
-
-            // Convert absolute path to app-relative path for security
-            string appBasePath = AppDomain.CurrentDomain.BaseDirectory;
-            string relativePath = _emailStoragePath.StartsWith(appBasePath)
-                ? _emailStoragePath.Substring(appBasePath.Length)
-                    .TrimStart(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar)
-                : Path.GetFileName(_emailStoragePath);
-
-            return Ok(
-                new
-                {
-                    status = "healthy",
-                    storagePath = relativePath,
-                    encryption = encryptionEnabled ? "enabled" : "disabled",
-                    validation = new
-                    {
-                        ipRequired = requireIpValidation,
-                        authRequired = requireAuthValidation
-                    },
-                    timestamp = DateTime.UtcNow
-                });
-        }
-
-        /// <summary>
-        /// Sanitizes a string for safe logging by removing newlines and control characters.
-        /// This prevents log injection attacks.
-        /// </summary>
-        private static string SanitizeForLog(string? input)
-        {
-            if (string.IsNullOrEmpty(input))
-            {
-                return string.Empty;
-            }
-
-            // Remove newlines, tabs, and control characters that could be used for log injection
-            var sanitized = LogSanitizationRegex().Replace(input, " ");
-
-            // Truncate to reasonable length to prevent log flooding
-            const int maxLogLength = 200;
-            if (sanitized.Length > maxLogLength)
-            {
-                sanitized = string.Concat(sanitized.AsSpan(0, maxLogLength), "...");
-            }
-
-            return sanitized;
-        }
-
-        /// <summary>
-        /// Masks an email address for logging to prevent PII exposure.
-        /// Example: "user@example.com" becomes "u***@e***.com"
-        /// </summary>
-        private static string MaskEmail(string? email)
-        {
-            if (string.IsNullOrEmpty(email))
-            {
-                return "[empty]";
-            }
-
-            var atIndex = email.IndexOf('@');
-            if (atIndex <= 0)
-            {
-                return "[invalid]";
-            }
-
-            var localPart = email[..atIndex];
-            var domainPart = email[(atIndex + 1)..];
-
-            // Mask local part: show first char + ***
-            var maskedLocal = localPart.Length > 0
-                ? $"{localPart[0]}***"
-                : "***";
-
-            // Mask domain: show first char + *** + TLD
-            var lastDotIndex = domainPart.LastIndexOf('.');
-            string maskedDomain;
-            if (lastDotIndex > 0)
-            {
-                var tld = domainPart[lastDotIndex..];
-                maskedDomain = $"{domainPart[0]}***{tld}";
-            }
-            else
-            {
-                maskedDomain = $"{domainPart[0]}***";
-            }
-
-            return $"{maskedLocal}@{maskedDomain}";
+            return Ok(_sendGridEmailService.GetHealthStatus());
         }
 
 #pragma warning disable CA1822 // Mark members as static - method accesses instance members via User property
